requirements.txt
annotated-types==0.7.0
cachetools==5.5.2
certifi==2025.1.31
charset-normalizer==3.4.1
google-api-core==2.24.1
google-auth==2.38.0
google-cloud-secret-manager==2.23.1
googleapis-common-protos==1.68.0
grpc-google-iam-v1==0.14.0
grpcio==1.70.0
grpcio-status==1.70.0
idna==3.10
proto-plus==1.26.0
protobuf==5.29.3
pulsar-client==3.6.0
pyasn1==0.6.1
pyasn1-modules==0.4.1
pydantic==2.10.6
pydantic-core==2.27.2
pydantic-settings==2.8.0
python-dotenv==1.0.1
requests==2.32.3
rsa==4.9
typing-extensions==4.12.2
urllib3==2.3.0
